+++ BaseCode/Overloads/OverloadedFuncs.h
@@ -7,7 +7,7 @@
 #include "../Exceptions/Exceptions.h"
 
 
-sf::Vector2f operator*(sf::Vector2f v1, int const& a);
+sf::Vector2f operator*(sf::Vector2f v1, float const& a);
 
 sf::Vector2f operator/(const sf::Vector2f v1, const sf::Vector2f v2);
